navigation component
ik heb de navigatie uit de layout.svelte gehaald
